+++ src/server.js
@@ -8,7 +8,7 @@ const colours = require('colors/safe')
 const CloudServer = require('./cloud-server.js')
 const fsUtil = require('./util.js')
 
-async function startServer ({ port, lockVars }) {
+async function startServer ({ port, lockVars, perMessageDeflate }) {
   const app = express()
   const cloudServer = new CloudServer({ lockVars })
 
@@ -26,7 +26,9 @@ async function startServer ({ port, lockVars }) {
     extensions: ['html', 'htm']
   }))
 
-  app.ws('/', cloudServer.handleWsConnection)
+  app.ws('/', cloudServer.handleWsConnection, {
+    wsOptions: { perMessageDeflate }
+  })
 
   app.use((req, res) => {
     res.status(404).sendFile(path.resolve(__dirname, '../static/404.html'))
@@ -36,8 +38,14 @@ async function startServer ({ port, lockVars }) {
     console.log(colours.green('I\'m now running your cloud server!'))
     console.log('You can access it...')
     console.log(`  • on your computer at ${colours.cyan(`ws://localhost:${port}/`)} (use this for testing)`)
-    console.log(`  • locally within your network at ${colours.blue(`ws://${await internalIp.v4()}:${port}/`)} (maybe)`)
-    console.log(`  • publicly at ${colours.blue(`ws://${await publicIp.v4()}:${port}/`)}, but ONLY if you've set up port forwarding on your router`)
+    const privateIp = await internalIp.v4().catch(() => null)
+    if (privateIp) {
+      console.log(`  • locally within your network at ${colours.blue(`ws://${privateIp}:${port}/`)} (maybe)`)
+    }
+    const publicIp = await publicIp.v4().catch(() => null)
+    if (publicIp) {
+      console.log(`  • publicly at ${colours.blue(`ws://${publicIp}:${port}/`)}, but ONLY if you've set up port forwarding on your router`)
+    }
     console.log(colours.yellow(`I'm also serving files from the static/ folder, which you can access in your browser at ${colours.blue(`http://localhost:${port}/`)}.`))
     console.log(colours.red('Press control+C to stop the server.'))
   })